add movie details to its own component
